Made changes for functional UI with bootstrap 5
This UI will still need modification, but both the sequencer and carrier
portions can now be browsed thanks to the use of nav-tabs and navbar. Badges
have also been fixed.

This will be updated soon so that the interface is more coherent, ideaslly
including the sequencer UI constructed within tabs.