Moved safe setting of nRST in relation to vreg_en
This will be completely safe once the pull-up is supported for the gpio. The setting of nRST low before claiming the vreg_en pin means that once this forces it low (with the pull-up) there will not be an issue, once this is added. However, for the current operation without the pull-up, the system is safe.