Add critical temperature monitoring
Monitoring of the 'ASIC' for a critical temperature has been added to the ASIC
adapter. When a critical temperature is detected, the regulators will be forced
off, after which human intervention will be needed to cycle the regulators to
recover the system.

Currently, the system is using the BME280 on-board for this, but eventually (once
the PCB supports it) this should be the direct in-ASIC sensors. For now, the
shroud that covers the ASIC also covers the BME280, and so the temperature it
reads is usually within a degree of the coolant.

Some modifications to the regulator and power cycling behaviour have been changed
so that disabling the regulators will not cause a crash.