Add controls for feedback cap and frame control
Adapter access functions as well as corresponding parameter tree elements and UI controls
have been added for preamp feedback capacitance, integration time in frames, and frame
time in cycles.

The adapter access functions are also designed to cache the values so that they can be
returned without querying the ASIC, which is the desired behaviour for all other access
functions in the future to decouple ASIC access load from the number of UI clients or
requests.