Add base serialiser config and more config bits
More bits can not be configured within the serialiser config class, and some additional access
functions have been added so that this can be done externally. Additionally, there is a base
configuration (differing from the default values on reset) for the serialiser config blocks
that has been determined through experimentation and guidance.

Additionally, the ASIC now returns a custom error if an attempt is made to read/write SPI while
the ASIC is known to be in reset.