Expose calibration test pattern settings for grid ID
The ASIC calibration test pattern can now be set as part of the ASIC class. This has
been extended to include the ability to set a specific single 4x4 grid intersection
of a division and sector so that it is easier to pinpoint specific parts of the image.
This has been exposed to javascript functions, but as of yet there are no UI elements.

There hava also been functions added for the test shift register internally, so that
patterns can be written out. In particular, the test_pattern_identify_sector_division()
function allows each 4x4 grid to be programmed with a unique ID in the pixel values,
which should be read out of the fast data. This also allows boundaries to be found
more easily. Note that it is only possible to program one sector at a time with this
method.