Start the system in a disabled state
This means that if the controller is booted while the target power supplies are
turned off, it no longer attempts to continue and breaks the bus. The user has
a chance to set up the equipment before clicking the 'Power Cycle Regulators'
button to actually establish communication.

This still does not stop the buses becoming borked if the button is clicked
while the target PCB is not powered.