Add prototype code for board connection monitoring
This should allow the system to check the presence of a powered PCB before attempting
to enable the regulators and contact devices (which currently borks the buses requiring
a reboot). The structure is now present, but the actual mechanism for testing presence
without breaking the buses had not been found.