Update temperature ajax to be async
This is the first modification to be made complex by having to separate logic
from outside the ajax request with logic from inside the response to make
them async-compatible. To do this, the firefly request loop must recover
the firefly ID from inside, as the callback time in
relation to the loop running outside cannot be gguaranteed. Therefore, the
context for the request, recovered using , must be used to inspect
the original url contents. This model can be re-used in other updates.

In addition, the chart processing logic that requires values to have been
updated within the response callback has now been
moved within the callback itself.