Update HTTP requests to use native fetch
This replaces the previous (JQuery) AJAX calls, but still operates asynchronously
thanks to the use of Promises. In some places, async functions have been used with
await to retain some synchronous behaviours, but only where necessary (and in such
a was as not to hold up the main thread).

The PUT and GET functions used are part of the to-be-standardised odin_control.js,
which has also been upgraded to allow the user of a timeout along with the request.