Sequences better segregated, and readout updated for progress
The readout sequence now makes use of the abort and progress reporting functionality
gained in the latest odin-sequencer update.

In addition, following the latest beamtime, the existing sequences as well as tests
have been separated out in a more logical fashion. The underlying support functions
are now separate to the tests, and some older 'legacy' support functions have been
separated out, and should not be used in the future (simply do not include the file).

The tests have also been added to the repository, but should not necessarrily be
kept here in the future.