Added ability to save a session name for test outputs
The session name will allow experiments to be grouped together, rather
than the previously hard-coded (in an inner sequence) name relating to
a specific test session. This way it can be either left (in which case
a default directory can be used) or set explicitly.

Assuming that the exports directory has been mounted to a non-volatile
location (as it would be to have modifiable sequences) this will
persist.