Added a default delay of 1ms between sector reads
This is a heavyhanded way of ensuring that the UI remains responsive while the SPI
data capture sequences run, or the HTTP requests are not answered. This is the
minimum delay that works, and it incurs a 20% time penalty.

Therefore this has been made a parameter of the readout function so that it can
be disabled if necessary.

A few more tweaks have also been added to the readout function(s) to document the
parameters and also force a 'dirty' abort when the abort has been triggered by using
an Exception, which means it will work even if the caller does not check the abort
flag. However, it does prevent the caller from performing cleanup actions.