Iniital additon of Munir thread-safe sequence DPDK control
The Munir adapter is used to control DPDK for fast data reception. This is operated as an
adapter on another instance of odin-control, and is therefore accessed via a proxy adapter.

The ProxyContext is then used to protect the asychronous calls to the proxy adapter and allow
the MunirProxyContext to be operated in synchronous sequences.

This is a dirty but working version. The system has brought the HEXITEC-MHz ASIC to data
output mode and controlled munir to store fast data via DPDK remotely.