Added Proxy Context for GPIB Controller
This allows GPIB instruments to be operated by an external device (RPi) running the GPIB
adapter on its own instance. As with Munir, this is asynchronous but protected so that it
can be operated via sequences.

Helper functions have been added so that certain useful functions for the existing
experimental setup are available for sequences. The path endpoints for 2510 and 2410 units
that are expected for this setup are found automatically and used as an ASIC bias and
peltier temperature controller respectiveley.

If these devices are not present, or the GPIB adapter is not present, the wider system will
still function.