Add ASIC register_controller instance with HMHz skeleton functions
A new directory for the loki-based carrier and ASIC code has been added
alongside the old one, and will eventually replace it entirely.

The adapter has been added, but is more standard than the previous
mercury one since it was complicated by attempts to create asynchronous
loops. This is now handled automatically by the LOKI carrier class, so
is not required.

The new asic class, in hexitec_mhz_asic.py, is very similar to the
implementation for BabyD, and uses the now standardised
register_controller class that builds in caching and easy access to
named fields. Some fields have already been added, but the rest of the
register map will need adding. Skeleton functions have been added for
the functions that were previously available in the old ASIC, class, so
that the same interface can be re-created, meaning minimal impact on
external code (besides it becoming far easier to alter settings in the
register map).

The new carrier class has not yet been committed, but will be derived
from the new LokiCarrier series of classes.