Update carrier firefly handling, adapter now runs
The adapter has been altered to the point where it can actually be
included without crashing. The parameter tree has been updated to
reflect modifications.

The channel mapping is now improved such that access functions (the
fireflies particularly currently) that act on lanes are called
referencing ASIC output channel numbers. For example, deactivating a
firefly output will use the ASIC channel number rather than a firefly
output number on a certain device. The mapping of ASIC channels to
firefly device and channel is now internal, and stored in a data
structure so that it can be easily modified.

Environment monitor has also been updated to include both fireflies.