Add device enable / system flow state machine
There is now a state machine implemented in a new task that will handle
the initialisation of the system on a step-by-step basis, with an
awareness of boards that are/aren't connected (using presence GPIO
inputs).

This is implemented as a standard state machine, except that there is a
'target' state. The machine will step through automatically and halt at
this state, or jump to it if it is a previous state.

The state jumps allowed via external functions are more limited, and
essentially allow the initialisation (or re-init) of the LOKI carrier,
power board, COB and ASIC in that order.

Basic config for AD7998 also added.