Added LTC2986 ASIC diode monitoring and ADC trips / HV
Added setup for the on-LOKI-carrier LTC2986, configuring it (if the COB
is present) to use the ASIC's on-fabric temperature diode. This is
untested but very similar to the original method. This requires an
update to the LOKI base class, which now provides more advanced initial
setup for the LTC device.

More derived information is now available from the ADC readings, like a
calculated HV voltage and named trips, both of which are now in the
parameter tree.

Some line names have been corrected since they included 'firefly'
despite not being related.

Lock handling has been improved.