Add High Voltage Bias Control Functions
The high voltage power supply can now be set for a specific target bias
voltage. This is possible by using one of two modes: 'auto' or 'manual'.

In 'manual' mode, the control voltage required is determined from
calibration / the data sheet conversion graph and assigned immediately.
This can also be set directly as an input control voltage if desired.

In 'auto' mode, the target bias voltage is associated with the voltage
monitor readback line, which is read via the ADC. This will dynamically
change the control voltage to reach its target. Currently untested due
to non-working ADC.

In both modes, there is readback of the current actual bias voltage
derived from the monitor voltage above, which relies on the ADC.

HV will not be enabled if the system fails to init the ADC or digital
potentiometer. If a control voltage setting has been saved to the
EEPROM, it will become active after the HV system inits. There is also
the option to save the current setting (once the user is happy) to the
EEPROM for next boot.