Allow LTC to use diode in single and differential mode on COB
First version of COB has TEMP1 and TEMP2 reversed (the LTC will only accept one polarity into its channels). Fortunately it is possible to read the diode in single ended mode using the lower channel (5), so long as channel 6 has been grounded with the solder bridge on either the power board or COB. Support for both has been included, with a config setting to switch between the two.