Calibration pattern auto apply and more patterns
The calibration pattern previously required the calling of a method
every time settings were changed to actually propagate them, which
became unintuitive in the UI. Now, every time a calibration pattern
setting is changed (whether or not it is relevant to the current mode),
the latest settings are synced to the ASIC. Realistically this does not
add much additional superfluous communication, and results in a system
that is more intuitive. However the user is still able to alter settings
for one calibration mode without changing to it immediately.

There are also slight changes to the parameter tree such that the UI can
scan for different modes more easily without knowing them, and two
additional asymmetrical patterns have been added (because previous
patterns were simple grids, making it impossible to use them to judge if
channels had been switched etc). While checking specific channels was
possible already by setting specific pixel or 4x4 grids, it is nice to
be able to get an immediate visual impression of if everything looks
correct or not.