Better re-init behaviour for ASIC setup
The ASIC re-initialisation behaviour previously did not actually power
cycle the regulators again or disable the ASIC. This has now been
corrected, so that any re-init of the ASIC should be exactly the same as
the first initialisation. This has been tested when attempting to sync
fast data channels, and prevents the user from having to disable the
regulators manually before clicking the button.