Fix serialiser config regmap inline with prototype
There was an issue with one serialiser configuration field causing its
default value in the ASIC not to be properly overwritten. This has now
been corrected.

The register mapping of the whole ASIC is now completely inline with the
prototype system after fast data has been started, bit-for-bit.

To make this comparison, a 'legacy' mode was added to the sequence for
the ASIC register dump so that the files are directly comparable.