Update default configuration settings
The fast data is now enabled in 'simple' mode by default, which doesn't
connect to the FireFlies (simply leaves all channels enabled). This
could be updated later when there is more confidence in the FireFly
control.

HV is now not enabled immediately after initialisation by default. This
can still be toggled in the UI, or be changed to True in the config so
that it comes up immediately for 'production' systems.