Add compatibility support for munir triggering of histogramming data
The histogramming software can now be trigged with the munir adapter.
These sequences are designed to operate without any changes to the proxy
context, and as such use minimal arguments. The num_batches argument is
repurposed for use specifying the number of frames per histogram.

Eventually, these arguments should be added to the munir adapter, as
well as the proxy context.