Read previously set HV bias target on start
If a bias has previously been set, and saved in the potentiometer for
the HV, it will now be read as an HV voltage and populated in the UI
from start (unless it is otherwise overridden). This makes the UI a
little less confusing, and also means you can tell what the HV is set to
before you turn it on or change it manually.