By default, init FireFlies properly
Previously, issues controlling FireFly select lines meant that it was
safer to just ignore them by default- just enabling them and not
establishing control over I2C. This works because by default, all
channels are enabled anyway. Whether the control software attempts more
comprehensive control depends on an environment variable set in the
config file.

Since the issues now seem to be resolved, the default setting is that
full control is enabled. The 'simple' mode can be enabled in a custom
configuration file if desired (if there are issues distinguishing the
two FireFly modules for example).