Add dew point as env sensor
The dew point is now calculated as another environment sensor, meaning
it just latches onto that monitoring thread. Since the idea is to
monitor the general case environment, I've used the power board
temperature sensor (MIC284 internal) with a fallback of LOKI's BME280.

For the humidity I had to use the BME280, since it is the only sensor
available.