Add dew point and critical temperature reactions
The dew point and temperature vs critical temperature are now checked at
strategic points during system init. If the limits are exceeded, the
system will stop the init process and return to the last stable state.
If dew point has been reached, HV will also always be disabled.

While in normal operation (ASIC_INIT) state, both are monitored
continuously. If tripped, the system will return to COB_INIT state with
HV disabled, and the ASIC disabled.

A status message has also been added to the state machine so that the UI
can display what is happening as the system steps through the init
process. This also reassures the user that the system has not locked up.